fixed integration between the work order details
