+++ client/src/components/work-orders/WorkOrderList.js
@@ -1,25 +1,12 @@
 import { useState, useEffect } from "react"
-import { getWorkOrders } from "../../managers/workOrderManager"
 import { WorkOrderCard } from "./WorkOrderCard"
 
-export const WorkOrderList = ({ setWoBikeId, loggedInUser }) => {
-  const [allWorkOrders, setAllWorkOrders] = useState([])
+export const WorkOrderList = ({
+  setWoBikeId,
+  allWorkOrders,
+  getAllWorkOrders,
+}) => {
   const [incompleteWorkOrders, setIncompleteWorkOrders] = useState([])
-
-  const getAllWorkOrders = () => {
-    if (loggedInUser?.isAdmin) {
-      getWorkOrders().then((woArray) => {
-        setAllWorkOrders(woArray)
-      })
-    } else {
-      getWorkOrders().then((allWoArray) => {
-        const userWorkOrders = allWoArray.filter(
-          (wo) => wo.bike.userId === loggedInUser?.id
-        )
-        setAllWorkOrders(userWorkOrders)
-      })
-    }
-  }
 
   useEffect(() => {
     getAllWorkOrders()
