+++ client/src/components/work-orders/WorkOrders.js
@@ -1,9 +1,26 @@
 import { useState } from "react"
 import BikeDetails from "../bikes/BikeDetails"
 import { WorkOrderList } from "./WorkOrderList"
+import { getWorkOrders } from "../../managers/workOrderManager"
 
 export const WorkOrders = ({ loggedInUser }) => {
   const [woBikeId, setWoBikeId] = useState(null)
+  const [allWorkOrders, setAllWorkOrders] = useState([])
+
+  const getAllWorkOrders = () => {
+    if (loggedInUser?.isAdmin) {
+      getWorkOrders().then((woArray) => {
+        setAllWorkOrders(woArray)
+      })
+    } else {
+      getWorkOrders().then((allWoArray) => {
+        const userWorkOrders = allWoArray.filter(
+          (wo) => wo.bike.userId === loggedInUser?.id
+        )
+        setAllWorkOrders(userWorkOrders)
+      })
+    }
+  }
 
   return (
     <div className="container">
@@ -11,11 +28,16 @@ export const WorkOrders = ({ loggedInUser }) => {
         <div className="col-sm-8">
           <WorkOrderList
             setWoBikeId={setWoBikeId}
-            loggedInUser={loggedInUser}
+            allWorkOrders={allWorkOrders}
+            getAllWorkOrders={getAllWorkOrders}
           />
         </div>
         <div className="col-sm-4">
-          <BikeDetails detailsBikeId={woBikeId} />
+          <BikeDetails
+            detailsBikeId={woBikeId}
+            user={loggedInUser}
+            getAllWorkOrders={getAllWorkOrders}
+          />
         </div>
       </div>
     </div>
