+++ client/src/managers/userManager.js
@@ -1,22 +1,18 @@
+import { fetchOptions } from "../helper"
+
+const apiUrl = "http://localhost:8088/users"
+
 export const getUserByEmail = (email) => {
-  return fetch(`http://localhost:8088/users?email=${email}`).then((res) =>
-    res.json()
-  )
+  return fetch(`${apiUrl}?email=${email}`).then((res) => res.json())
 }
 export const getUserById = (id) => {
-  return fetch(`http://localhost:8088/users/${id}`).then((res) => res.json())
+  return fetch(`${apiUrl}/${id}`).then((res) => res.json())
 }
 
 export const getUsers = () => {
-  return fetch(`http://localhost:8088/users`).then((res) => res.json())
+  return fetch(`${apiUrl}`).then((res) => res.json())
 }
 
 export const createUser = (user) => {
-  return fetch("http://localhost:8088/users", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(user),
-  }).then((res) => res.json())
+  return fetch(apiUrl, fetchOptions("POST", user)).then((res) => res.json())
 }
